feat(project-init): MethodArgumentNotValidException 에러를 핸들링하기 위한 invalidRequestException 메서드 설계 #26
